docshell/test/chrome/test_bug454235.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=454235
-->
<window title="Mozilla Bug 454235"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=454235"
     target="_blank">Mozilla Bug 454235</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[

    /** Test for Bug 454235 **/
SimpleTest.waitForExplicitFinish();

SimpleTest.waitForFocus(doTest);

function doTest() {
  var hiddenBrowser = document.getElementById("hiddenBrowser");

  hiddenBrowser.contentWindow.focus();
  ok(!hiddenBrowser.contentDocument.hasFocus(), "hidden browser is unfocusable");

  SimpleTest.finish();
}



  ]]></script>
  <box flex="1" style="visibility: hidden; border:5px black solid">
    <browser style="border:5px blue solid" id="hiddenBrowser" src="bug454235-subframe.xhtml"/>
  </box>
</window>
